create and list user and New DbContext(DB) For Users

+++ HelpDesk.MVC/Controllers/AdminController.cs
@@ -13,12 +13,15 @@ using HelpDesk.Domain.Core.Categories;
 using HelpDesk.InfraStructures.DataAccess.Common;
 using HelpDesk.MVC.Models.Articles;
 using HelpDesk.MVC.Models.categories;
-using MD.PersianDateTime;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using HelpDesk.Domain.Core.Images;
 using static System.Net.Mime.MediaTypeNames;
+using HelpDesk.MVC.Service;
+using MD.PersianDateTime.Standard;
+using Microsoft.AspNetCore.Identity;
+using HelpDesk.MVC.Models.Users;
 
 namespace HelpDesk.MVC.Controllers
 {
@@ -27,13 +30,17 @@ namespace HelpDesk.MVC.Controllers
         private readonly ICategoryRepository categoryRepository;
         private readonly IArticleRepository articleRepository;
         private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IUploadFileRepository uploadFileRepository;
         private readonly IimageRepository imageRepository;
-        public AdminController(IimageRepository imageRepository,ICategoryRepository categoryRepository, IArticleRepository articleRepository,IHostingEnvironment hostingEnvironment)
+        private readonly UserManager<ApplicationUsers> userManager;
+        public AdminController(IimageRepository imageRepository, ICategoryRepository categoryRepository, IArticleRepository articleRepository, IUploadFileRepository uploadFileRepository, IHostingEnvironment hostingEnvironment, UserManager<ApplicationUsers> userManager)
         {
             this.categoryRepository = categoryRepository;
             this.articleRepository = articleRepository;
             this._hostingEnvironment = hostingEnvironment;
+            this.uploadFileRepository = uploadFileRepository;
             this.imageRepository = imageRepository;
+            this.userManager = userManager;
         }
         public IActionResult Index()
         {
@@ -60,8 +67,8 @@ namespace HelpDesk.MVC.Controllers
         public IActionResult ListCat()
         {
             var cat = categoryRepository.GetAll().ToList();
-           // ViewData["ArticlesCount"] = articleRepository.GetArticlesCountInCategory()
-           return View(cat);
+            // ViewData["ArticlesCount"] = articleRepository.GetArticlesCountInCategory()
+            return View(cat);
         }
         public IActionResult ListArticle()
         {
@@ -118,6 +125,7 @@ namespace HelpDesk.MVC.Controllers
             return View(cat);
         }
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> AddArticle(AddNewArticleGetViewModel model)
         {
             if (ModelState.IsValid)
@@ -130,9 +138,25 @@ namespace HelpDesk.MVC.Controllers
                     Body = model.Body,
                     PublishDate = persiandate.ToString(),
                     Status = ArticleStatus.Publish,
-                    CategoryId=model.SelectedCat
+                    CategoryId = model.SelectedCat
                 };
-                if ((model?.Image?.Length > 0) && ((model?.Image?.ContentType=="image/jpeg") || (model?.Image?.ContentType == "image/jpg")))
+                //if ((model?.Image?.Length > 0) && ((model?.Image?.ContentType=="image/jpeg") || (model?.Image?.ContentType == "image/jpg")))
+                //{
+
+                //    string newFileName = uploadFileRepository.UplaodFile(model.Image, "\\Images\\");
+                //    article.Image = @"~/images/" + newFileName;
+                //}
+                //if ((model?.Video?.Length > 0) && (model?.Video?.ContentType == "video/mp4") )
+                //{
+                //    string newFileName = uploadFileRepository.UplaodFile(model.Video, "\\video\\");
+                //    article.Video = @"~/video/" + newFileName;
+                //}
+                //if ((model?.PDF?.Length > 0) && (model?.PDF?.ContentType == "application/pdf"))
+                //{
+                //    string newFileName = uploadFileRepository.UplaodFile(model.PDF, "\\pdf\\");
+                //    article.PDF = @"~/pdf/" + newFileName;
+                //}
+                if ((model?.Image?.Length > 0) && ((model?.Image?.ContentType == "image/jpeg") || (model?.Image?.ContentType == "image/jpg")))
                 {
                     string path_root = _hostingEnvironment.WebRootPath;
                     string path_to_image = path_root + "\\Images\\" + model.Image.FileName;
@@ -142,7 +166,7 @@ namespace HelpDesk.MVC.Controllers
                     }
                     article.Image = @"~/images/" + model.Image.FileName;
                 }
-                if ((model?.Video?.Length > 0) && (model?.Video?.ContentType == "video/mp4") )
+                if ((model?.Video?.Length > 0) && (model?.Video?.ContentType == "video/mp4"))
                 {
                     string path_root = _hostingEnvironment.WebRootPath;
                     string path_to_video = path_root + "\\Video\\" + model.Video.FileName;
@@ -163,9 +187,15 @@ namespace HelpDesk.MVC.Controllers
                     article.PDF = @"~/pdf/" + model.PDF.FileName;
                 }
 
+
                 articleRepository.Add(article);
+                return RedirectToAction(nameof(ListArticle));
+
             }
-            return RedirectToAction(nameof(ListArticle));
+            else
+            {
+                return RedirectToAction(nameof(ListArticle));
+            }
         }
         public IActionResult DeleteArticle(int id)
         {
@@ -202,38 +232,47 @@ namespace HelpDesk.MVC.Controllers
                 return View(displayArticleCategory);
             }
         }
+        [ValidateAntiForgeryToken]
         [HttpPost]
         public async Task<IActionResult> EditArticle(AddNewArticleGetViewModel displayArticle, Article article)
         {
-            var oldImagePath = displayArticle.ImagePath.Remove(0,9);
-            article.CategoryId = displayArticle.SelectedCat;
-            if ((displayArticle?.Image?.Length > 0) && ((displayArticle?.Image?.ContentType == "image/jpeg") || (displayArticle?.Image?.ContentType == "image/jpg")))
+            if (ModelState.IsValid)
             {
-                string path_root = _hostingEnvironment.WebRootPath;
-                string path_Old_Image = path_root + "\\Images\\" + oldImagePath;
-                string path_to_image = path_root + "\\Images\\" + displayArticle.Image.FileName;
-                using (var stream = new FileStream(path_to_image, FileMode.Create))
+                var oldImagePath = displayArticle.ImagePath.Remove(0, 9);
+                article.CategoryId = displayArticle.SelectedCat;
+                if ((displayArticle?.Image?.Length > 0) && ((displayArticle?.Image?.ContentType == "image/jpeg") || (displayArticle?.Image?.ContentType == "image/jpg")))
                 {
-                    await displayArticle.Image.CopyToAsync(stream);
-                }
-                article.Image = @"~/images/" + displayArticle.Image.FileName;
-                System.IO.File.Delete(path_Old_Image);
 
-                articleRepository.Update(article);
-                return RedirectToAction(nameof(ListArticle));
+                    string path_root = _hostingEnvironment.WebRootPath;
+                    string path_Old_Image = path_root + "\\Images\\" + oldImagePath;
+                    string path_to_image = path_root + "\\Images\\" + displayArticle.Image.FileName;
+                    using (var stream = new FileStream(path_to_image, FileMode.Create))
+                    {
+                        await displayArticle.Image.CopyToAsync(stream);
+                    }
+                    article.Image = @"~/images/" + displayArticle.Image.FileName;
+                    System.IO.File.Delete(path_Old_Image);
+
+                    articleRepository.Update(article);
+                    return RedirectToAction(nameof(ListArticle));
+                }
+                else
+                {
+                    article.Image = displayArticle.ImagePath;
+                    articleRepository.Update(article);
+                    return RedirectToAction(nameof(ListArticle));
+
+                }
             }
             else
             {
-                article.Image = displayArticle.ImagePath;
-                articleRepository.Update(article);
-                return RedirectToAction(nameof(ListArticle));
-
+                return View(article);
             }
         }
         public IActionResult ArticleDetails(int id)
         {
             var article = articleRepository.Get(id);
-            
+
             if (article == null)
             {
                 return BadRequest();
@@ -241,7 +280,7 @@ namespace HelpDesk.MVC.Controllers
             else
             {
                 Category category = new Category();
-                category= categoryRepository.Get(article.CategoryId);
+                category = categoryRepository.Get(article.CategoryId);
                 ViewData["Category"] = category.Name;
                 return View(article);
             }
@@ -276,11 +315,12 @@ namespace HelpDesk.MVC.Controllers
             }
         }
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Upload2(IFormFile file, Article article)
         {
             var uploads = Path.Combine(_hostingEnvironment.WebRootPath, "Images\\Multi");
             int i = HttpContext.Request.Headers.Count;
-           var request = 1;
+            var request = 1;
             if (request == 1)
             {
                 if (file.Length > 0)
@@ -302,12 +342,14 @@ namespace HelpDesk.MVC.Controllers
             return RedirectToAction("Index");
         }
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public void Upload2222(string name, int request)
         {
             var name2 = name;
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Upload(IFormFile file, Article article, string name, int request)
         {
             var request1 = request;
@@ -337,6 +379,77 @@ namespace HelpDesk.MVC.Controllers
 
             }
             return RedirectToAction("Index");
+        }
+
+        public IActionResult UserList()
+        {
+                var userlist = userManager.Users.Take(50).ToList();
+                return View(userlist);
+
+        }
+        public IActionResult CreateUser()
+        {
+            return View();
+        }
+        public async Task<IActionResult> UploadFile(IFormFile files)
+        {
+            //"upload\\userimage\\normalimage\\"
+            //"\\upload\\userimage\\thumbnailimage\\"
+            string newFileName = uploadFileRepository.UplaodFile(files, "\\Images\\Profile\\");
+            string path_root = _hostingEnvironment.WebRootPath;
+            string path_to_video = path_root + "\\Images\\Profile\\" + newFileName;
+            using (var stream = new FileStream(path_to_video, FileMode.Create))
+            {
+                await files.CopyToAsync(stream);
+            }
+
+            return Json(new { status = "success", message = "تصویر با موفقیت آپلود شد.", imagename = newFileName });
+
+        }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public  IActionResult CreateUser(UserViewModel model, string imagename)
+        {
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    if (imagename == null)
+                    {
+                        model.UserImage = "DefaultProfile.jpg";
+                    }
+                    else
+                    {
+                        model.UserImage = imagename;
+                    }
+
+                    ApplicationUsers user = new ApplicationUsers
+                    {
+                        FirstName = model.FirstName,
+                        LastName = model.LastName,
+                        PhoneNumber = model.PhoneNumber,
+                        UserName = model.UserName,
+                        Email = model.Email,
+                        
+                        Gender = model.gender,
+                        BirthDate = model.BirthDayDate,
+                        ProfileImage = model.UserImage
+                    };
+
+                    var result =  userManager.CreateAsync(user, model.Password).Result;
+                    if (result.Succeeded)
+                    {
+                        return RedirectToAction(nameof(Index));
+                    }
+
+                }
+                catch
+                {
+                    throw;
+                }
+            }
+            ViewBag.ViewTitle = "فرم ایجاد کاربر";
+            return View(model);
         }
